+++ hadoop-tools/hadoop-azure/src/main/java/org/apache/hadoop/fs/azurebfs/services/AbfsInputPolicy.java
@@ -1,3 +1,21 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.apache.hadoop.fs.azurebfs.services;
 
 import java.util.Locale;
@@ -19,8 +37,6 @@ public enum AbfsInputPolicy {
 
   SEQUENTIAL(FS_OPTION_OPENFILE_READ_POLICY_SEQUENTIAL),
   RANDOM(FS_OPTION_OPENFILE_READ_POLICY_RANDOM),
-  VECTORED(FS_OPTION_OPENFILE_READ_POLICY_VECTOR),
-  LAYOUT("layout"),
   ADAPTIVE(FS_OPTION_OPENFILE_READ_POLICY_ADAPTIVE);
 
   private final String policy;
@@ -38,11 +54,8 @@ public enum AbfsInputPolicy {
     return policy;
   }
 
-  public static AbfsInputPolicy getPolicy(String name, boolean isLayoutPresent) {
+  public static AbfsInputPolicy getPolicy(String name) {
     String trimmed = name.trim().toLowerCase(Locale.ENGLISH);
-    if (isLayoutPresent) {
-      return LAYOUT;
-    }
     switch (trimmed) {
     // all these options currently map to random IO.
     case FS_OPTION_OPENFILE_READ_POLICY_HBASE:
